services/api_gateway/requirements.txt
# API Gateway Dependencies
# Security Triage System - API Gateway Service

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
